Refactor main routes to their own file in routes folder.
