+++ assets/sass/funky_world_cup_light.scss
@@ -43,6 +43,3 @@
 
 // Nucleo Icons
 @import "lbd/partial-nucleo-icons";
-
-// Nucleo Icons
-@import "lbd/partial-nucleo-icons";
